+++ api/Controllers/User/Websites/DeleteController.cs
@@ -16,9 +16,6 @@ namespace FeChat.Controllers.User.Websites {
     // Used Mvc to get the Controller feature
     using Microsoft.AspNetCore.Mvc;
 
-    // Use the Authentication feature to get the access token
-    using Microsoft.AspNetCore.Authentication;
-
     // Use the Cors feature to control the access
     using Microsoft.AspNetCore.Cors;
 
@@ -43,6 +40,22 @@ namespace FeChat.Controllers.User.Websites {
     public class DeleteController: Controller {
 
         /// <summary>
+        /// Container for app's configuration
+        /// </summary>
+        private readonly IConfiguration _configuration;
+
+        /// <summary>
+        /// Constructor for this controller
+        /// </summary>
+        /// <param name="configuration">App configuration</param>
+        public DeleteController(IConfiguration configuration) {
+
+            // Add configuration to the container
+            _configuration = configuration;
+
+        }
+
+        /// <summary>
         /// Delete a website
         /// </summary>
         /// <param name="websiteId">Contains the website's ID</param>
@@ -51,8 +64,18 @@ namespace FeChat.Controllers.User.Websites {
         /// <returns>Message about the website status</returns>
         [HttpDelete("{websiteId}")]
         [EnableCors("AllowOrigin")]
-        [IgnoreAntiforgeryToken]
         public async Task<IActionResult> DeleteWebsite(int websiteId, Member memberInfo, IWebsitesRepository websitesRepository) {
+
+            // Verify if antiforgery is valid
+            if ( await new Antiforgery(HttpContext, _configuration).Validate() == false ) {
+
+                // Return error response
+                return new JsonResult(new {
+                    success = false,
+                    message = new Strings().Get("InvalidCsrfToken")
+                });
+
+            }
 
             // Delete a member
             ResponseDto<bool> deleteWebsite = await websitesRepository.DeleteWebsiteAsync(websiteId, memberInfo.Info!.MemberId);
